# Nginx: Refactored "ngx_autoblock" to reduce cpu consumption