This PR addresses a CPU consumption issues I have with "`ngx_autoblock.php`". The script dominates the top process list, it consumes more CPU than all other processes and it runs for almost 30 seconds (every minute).

### Environment:
* Latest Opnsense on Intel quad core.
* ~**955** banned IP addresses (collected over time in ~1 year)
* Average load 1.3 (prior) and 0.8 (after applying the changes)

### Performance:
**Original "`ngx_autoblock.php`" (24 seconds +- 3)**: 
```.sh
time /usr/local/opnsense/scripts/nginx/ngx_autoblock.php
7.557u 16.446s 0:24.03 99.8%	4390+1547k 0+0io 0pf+0w
```
**Updated "`ngx_autoblock.php`"**:

| Condition                      | Time results  |
| ---------------------  |:-------------:|
| Regular schedule with no changes | 0.690u 0.839s 0:01.55 |
| Every 10 minutes, no changes | 4.047u 5.147s 0:09.31 |
| Adding one IP from ban file to PF | 4.932u 6.848s 0:11.83 |
| Adding all IPs to PF | 3.861u 5.614s 0:09.84 |

The updated version is ~12 times faster for the regular schedule and still ~3 times faster for the full run "every 10 minutes".
**Note**: "*Adding one IP from ban file to PF*" is slower than "*Adding all IPs to PF*", since adding an IP from ban file updates the model.

### Hotspots
The profiler shows 2 hotspots:
* ~70% of the time is spent to execute "`pfctl ... -T add $singleIP`".
* ~30% of the time is spent for config loading & model parsing.

### Changes
* Monitoring sources for changes to see if an action is required.
  This saves time in both hotspots, model parsing and `pfctl` operations.
* Every 10 minutes the script runs regardless of the monitoring results to avoid missing changes e.g. done directly via `pfctl`.
*  `pfctl` is invoked with batches of multiple IPs. 
    This reduces the 70% hotspot to almost nothing.
* `pfctl` is used to read the contents of the table and incrementally sync the table to the desired state instead of always populating the full list of IPs.

### Fixes
- `$is_ten_minutes` triggered always except every 10 minutes. 
   Intended was likely to trigger every 10 minutes.
- IPs are not just added but also removed. Fixes possible race conditions between UI and script-invocation that may perfectly well happen due to the long runtime of the script.