@jkellerer 2020-04-10
After uninstalling the "debug" package, I have updated performance results (my original issue with "ngx_autoblock" was before installing them, but I have also applied changes from "[#4019](https://github.com/opnsense/core/pull/4019)" which was part of the performance problem):

| Condition                      | Time results  |
| ---------------------  |--------------:|
| Original "ngx_autoblock" | 4.041u 12.636s 0:16.55 |
| Regular schedule with no changes | 0.126u 0.071s 0:00.19 |
| Every 10 minutes, no changes | 0.479u 0.213s 0:00.71 |
| Adding all IPs to PF | 0.622u 0.236s 0:00.87 |

Worsts case 16x, best case 80x quicker.