@fabianfrz 2020-07-06
Thanks for the changes. I will try to test them as soon as I have time. 18x-80x faster is awesome.

I never had that load on my system as it is a home network so thanks for your contribution in advance.