@jkellerer 2020-07-07
Hi @fabianfrz , thanks for the plugin and thanks for reviewing my contribution.

The changes are running for quite a while now in my system and it has not produced any issues after my latest commit to address concurrency. I've had concurrency issues with the unmodified code as well, but much less often, probably due to different timing. ngx_autoblock is called twice sometimes and one invocation may have quit with a PHP error prior to the latest commit in this PR.

One other note: On a system that is public reachable the list of blocked IPs grows and grows. When looking at the blocked counts, it is quite effective (more blocks than the common block-lists I've installed). Eventually it would make sense to expire banned IPs after some time to avoid having unbound growth. The changes I proposed would offer a simple path to implement expiration since the code now actively manages the block list (not just adding entries). Once this is PR is accepted, I'd be happy to propose this as a FR case and also provide a new PR for it if this is ok for you.