@fabianfrz 2020-07-07
@jkellerer I can remember that someone asked for that in the past so I think it is a good idea. In my opinion it should be configurable so if for example if you configure 4 days than it will unblock them after 4 days unless there is a new incident from that IP in mean time. I developed that feature in that way because with the alias you can block all of those hosts - for example they might be spammers to as the host is likely compromised so they also cannot reach TCP/25 for example.