@jkellerer 2020-07-07
I fully agree to what you wrote, that's also how I would have done it. Once this is reviewed, I'll create a FR case to follow up on expiration and probably also a WIP-PR to have some more basis to discus on.