@fabianfrz 2020-07-28
I tested it successfully so @fichtner @AdSchellevis I can merge that but I have to add a release note afterwards and a version bump.