@8191 2021-01-12
@AdSchellevis Sorry to be so persistent here, but #2163 constantly leads to availability issues due to performance bottle-necks with one of my OPNsense instances. Is there anything needed so that this PR could be merged?
The PR was approved by the plugin maintainer (@fabianfrz) and intensively tested by the author (see PR description).