@AdSchellevis 2021-01-12
@8191 we're only mildly involved with most of these plugins, usually do a quick scan if it doesn't seem to break core components in some way (which it doesn't seem to do).  If @fabianfrz is ok with it, he can merge it. The coding quality doesn't seem to be totally intact by the way, it misses some curly braces if I saw correctly. Most of these will be fixed by a style sweep, but it's good to pay attention upfront.